refactor: wire recommendation-aware refresh
