+++ src/codex_usage_tracker/server/tach.domain.toml
@@ -7,5 +7,6 @@ depends_on = [
     "//codex_usage_tracker.diagnostics",
     "//codex_usage_tracker.pricing",
     "//codex_usage_tracker.reports",
+    "//codex_usage_tracker.recommendation_engine",
     "//codex_usage_tracker.store",
 ]
